Proposal: add a `style` property to simplify CSS one-off injection
Even if CSS belongs to the CSS world, having the ability to confine
all component related things in one file is still somehow handy
and often demanded/adopted by the Web development community.

I have tested this with the project _herey_, and I think it might
give wickedElements an extra reason to feel wicked.

For those not needing CSS in components this won't change anything,
but for those looking to ship components as a whole in one file
this is a core game-changer.

Question to the users, if anyone: is `style` the right propety to use?
Would this interfere with your component definition/life cycle?

I've thought some might use `this.style = this.el.style` in some component,
and I want to be sure this is backward compatible friendly.

Thanks to whoever would like to express an opinion.